Added changes after a code review.

+++ src/main/java/com/example/animalshelter/controller/AnimalController.java
@@ -30,7 +30,7 @@ public class AnimalController {
     }
 
     @GetMapping("/{id}")
-    public Animal one(@PathVariable Long id) {
+    public Animal one(@PathVariable int id) {
         return animalService.findOne(id);
     }
 }
