+++ src/main/java/com/example/animalshelter/service/AnimalServiceImpl.java
@@ -32,8 +32,8 @@ public class AnimalServiceImpl implements AnimalService {
     }
 
     @Override
-    public Animal findOne(Long id) {
-        return animalRepository.findOne(id)
+    public Animal findOne(int id) {
+        return animalRepository.findOne((long) id)
                 .orElseThrow(() -> new AnimalNotFoundException(id));
     }
 }
